requirements.txt


# Updated requirements.txt - Fixed bcrypt issues
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.3
redis==5.0.1
python-jose[cryptography]==3.4.0
bcrypt==4.1.2
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0